Update Lqr library to Git head. CCBUGS: 149485
svn path=/trunk/extragear/graphics/digikam/; revision=964658

+++ imageplugins/contentawareresizing/lqr/lqr_energy.c
@@ -619,7 +619,7 @@ lqr_carver_get_energy(LqrCarver * r, gint orientation)
       for (x = 0; x < w; x++)
         {
           data = orientation == 0 ? r->raw[y][x] : r->raw[x][y];
-          nrg = tanh(r->en[data]);
+          nrg = tanhf(r->en[data]);
           nrg_max = MAX (nrg_max, nrg);
           nrg_min = MIN (nrg_min, nrg);
           nrg_buffer[z0++] = nrg;
